feat: Move tokenizer things into the meilidb-tokenizer workspace
