+++ Cargo.toml
@@ -2,4 +2,5 @@
 members = [
     "meilidb",
     "meilidb-core",
+    "meilidb-tokenizer",
 ]
